fixed style to feed page
