+++ src/client/Pages/Feed.jsx
@@ -3,8 +3,8 @@ import React from "react";
 import Layout from "../Components/Layout";
 import { makeStyles } from "@material-ui/core/styles";
 import { Link as BrowserLink } from "@reach/router";
-import Card from "../Components/Card";
-import TextField from "../Components/TextField";
+import Post from "../Components/Post";
+import PostForm from "../Components/PostForm";
 
 const useStyles = makeStyles(theme => ({}));
 
@@ -19,8 +19,10 @@ export default function Feed({ onSubmit }) {
           justify="flex-end"
           alignItems="center"
           direction="column">
-          <TextField onSubmit={onSubmit} />
-          <Card />
+          <PostForm onSubmit={onSubmit} />
+          <Post />
+          <Post />
+          <Post />
         </Grid>
       </Container>
     </Layout>
